Adds new versions of Sapnish manuals uploaded by Mario to the server
Newer versions of the manual "Primeros pasos con Maxima" and
"Cálculo Numérico con Maxima" were added by Mario in February
of 2015, but they were never committed in this repository and
the documentation web page was not updated: it still pointed
to an old version of the second manual and the dates and authors
wer outdated too. This commit fixes that.