The snappy app for the Release.
Unlike traditional packages snappy apps package all that is needed in
order to make the application run. Which means that wxMaxima is packaged
with wxWidgets, maxima, sbcl and clisp. In theory we could make a separate
package for maxima. But in this case the package for maxima and the one
for wxMaxima each would install a separate maxima on the system.