Change the format of release dates on the English pages
This uses "Month Date, Year" instead of "Month, Date Year" or
"Month, Date, Year".  This is for consistency and because the
placement of the comma in the first form is the common way to
write dates in American English.  I don't think the other forms
are common anywhere (based only on some brief searches).